+++ include/ftk/mesh/simplicial_xgc_3dff_mesh.hh
@@ -8,7 +8,7 @@
 namespace ftk {
 
 /// legacy code: a magnetic field folowing mesh with simplicial elements
-template <typename I=int, typename F=double>
+template <typename I=int128_t, typename F=double>
 struct simplicial_xgc_3dff_mesh : public simplicial_xgc_3d_mesh<I, F> {
   simplicial_xgc_3dff_mesh(std::shared_ptr<simplicial_xgc_2d_mesh<I, F>> m2_, int nphi, int iphi=1, int vphi=1) 
     : simplicial_xgc_3d_mesh<I, F>(m2_, nphi, iphi, vphi) {}
